minor code improvements
---

there were a total of two fixes that DrMutoPackTool needed:
1. better cpp filenames.
previously they were all named as sequential number, from 1 all the way to 9.
this is no longer the case. instead, only 1.cpp and 1.hpp get to be named like this, while the rest had to be renamed.
their cpp filenames all start with 1, but following that are names that better reflect the functions that take place inside those files.

2. attempt to make "input file path"variable an UTF8 string, somehow.
i don't know how to explain this, but the short of it is it cropped up new issues of their own, see fixes 3-4 below for more details.

3. fix how folders are constructed.
it's a minor fix that hopefully no one will notice but due to how said path var is now an UTF8 string, that meant creating folders out-of-nowhere and on-the-fly no longer worked the way it used to, as code for doing that relied on removing the last char of said path.
this is no longer the case; all folders that packtool finds as the TOC is read is created without the above oddity if user chooses to extract the bigfile on their own.

4. fix how certain files are extracted from bigfiles.
this was born of an oversight over how file extraction code actually worked.
say that a certain file ended up being "aligned" to an additional 2048 bytes somehow, in some cases it would end up omitting said bytes without even blinking.
no longer the case here; for as long as the file size value from the TOC is known, all files are fully extracted without any incomplete data left out from the bigfiles.